move code remove dead code
